+++ __tests__/v2/admin_group_edit_page.test.js
@@ -16,7 +16,8 @@ vi.mock('$app/stores', () => {
 				],
 				group: {
 					name: 'test',
-					user_ids: [1]
+					user_ids: [1],
+					viewer_paths: []
 				}
 			}
 		})