+++ src/routes/+layout.svelte
@@ -3,6 +3,7 @@
 	import { afterNavigate, invalidateAll } from '$app/navigation';
 	import { page } from '$app/stores';
 	import { navigating } from '$app/stores';
+	import { env } from '$env/dynamic/public';
 	import { reloadVersionedPage } from '$lib/common/selected_api_version';
 	import { onMount } from 'svelte';
 
@@ -198,6 +199,9 @@
 							<li><a class="dropdown-item" href="/profile">My profile</a></li>
 							{#if $page.data.runnerBackend !== 'local' && $page.data.runnerBackend !== 'local_experimental'}
 								<li><a class="dropdown-item" href="/settings">My settings</a></li>
+							{/if}
+							{#if env.PUBLIC_FRACTAL_VIZARR_VIEWER_URL}
+								<li><a class="dropdown-item" href="/viewer-paths">Viewer paths</a></li>
 							{/if}
 							<li><a class="dropdown-item" href="/auth/logout">Logout</a></li>
 						</ul>